+++ json_serializable/test/src/checked_test_input.dart
@@ -4,7 +4,10 @@ part of '_json_serializable_test_input.dart';
 WithANonCtorGetterChecked _$WithANonCtorGetterCheckedFromJson(
     Map<String, dynamic> json) {
   return $checkedNew('WithANonCtorGetterChecked', json, () {
-    $checkKeys(json, allowedKeys: ['items']);
+    $checkKeys(json,
+        allowedKeys: const ['items'],
+        requiredKeys: const ['items'],
+        disallowNullValues: const ['items']);
     var val = WithANonCtorGetterChecked($checkedConvert(
         json, 'items', (v) => (v as List)?.map((e) => e as String)?.toList()));
     return val;
@@ -13,6 +16,7 @@ WithANonCtorGetterChecked _$WithANonCtorGetterCheckedFromJson(
 ''', checked: true)
 @JsonSerializable(disallowUnrecognizedKeys: true, createToJson: false)
 class WithANonCtorGetterChecked {
+  @JsonKey(required: true, disallowNullValue: true)
   final List<String> items;
   int get legth => items.length;
 
@@ -21,13 +25,17 @@ class WithANonCtorGetterChecked {
 
 @ShouldGenerate(r'''
 WithANonCtorGetter _$WithANonCtorGetterFromJson(Map<String, dynamic> json) {
-  $checkKeys(json, allowedKeys: ['items']);
+  $checkKeys(json,
+      allowedKeys: const ['items'],
+      requiredKeys: const ['items'],
+      disallowNullValues: const ['items']);
   return WithANonCtorGetter(
       (json['items'] as List)?.map((e) => e as String)?.toList());
 }
 ''')
 @JsonSerializable(disallowUnrecognizedKeys: true, createToJson: false)
 class WithANonCtorGetter {
+  @JsonKey(required: true, disallowNullValue: true)
   final List<String> items;
   int get legth => items.length;